refactor(ui): decompose FileExplorer.tsx into 8 focused modules
